pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-excerpt"
version = "0.1.0"
description = "Multi-bot Telegram → PRD generator with Cloud Run + Firestore backend"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Oreste" }]
keywords = ["telegram", "bot", "llm", "prd", "openrouter", "firestore"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]
dependencies = [
    "python-telegram-bot>=21.6,<22",
    "openai>=1.51,<2",
    "google-cloud-firestore>=2.19,<3",
    "fastapi>=0.115,<0.120",
    "uvicorn[standard]>=0.32,<0.40",
    "httpx>=0.27,<0.29",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",
    "structlog>=24.4,<26",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<0.26",
    "pytest-cov>=5.0,<7",
    "respx>=0.21,<0.23",
    "ruff>=0.7,<0.14",
    "mypy>=1.13,<2",
    "pip-audit>=2.7,<3",
    "pre-commit>=4.0,<5",
]

[project.scripts]
telegram-excerpt = "telegram_excerpt.__main__:main"

[project.urls]
Homepage = "https://github.com/oreste/telegram-excerpt"
Issues = "https://github.com/oreste/telegram-excerpt/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "PL",   # pylint
    "PT",   # flake8-pytest-style
    "D",    # pydocstyle
    "ANN",  # flake8-annotations
    "ASYNC", # flake8-async
    "S",    # bandit (security)
]
ignore = [
    "D100",  # Missing docstring in public module (too noisy)
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "ANN401", # Dynamically typed expressions (Any) allowed when needed
    "PLR0913", # Too many arguments
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S101", "ANN", "PLR2004"]
"src/telegram_excerpt/__main__.py" = ["D"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["google.cloud.firestore", "google.cloud.firestore_v1.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --cov=telegram_excerpt --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["src/telegram_excerpt"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
